Added FilterInformationTypeRequest which extends AbstractFilterRequest. Removed old class FilterRequest Made adjustments in the controllerclass Fixed existing test of InformationTypeSpecifications
