+++ data-catalog-backend-test/data-catalog-backend-test-component/src/test/java/no/nav/data/catalog/backend/test/component/informationtype/InformationTypeSpecificationTest.java
@@ -8,20 +8,14 @@ import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 
-import no.nav.data.catalog.backend.app.codelist.CodelistRepository;
-import no.nav.data.catalog.backend.app.informationtype.FilterRequest;
-import no.nav.data.catalog.backend.test.component.ComponentTestConfig;
+import no.nav.data.catalog.backend.app.informationtype.FilterInformationTypeRequest;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.Collections;
 import java.util.Map;
@@ -35,7 +29,7 @@ public class InformationTypeSpecificationTest {
 	public void mapFromQuery_withEmptyQuery_willReturnDefaultRestResponsePage() {
 		Map<String, String> queryMap = Collections.emptyMap();
 
-		FilterRequest filterRequest = new FilterRequest().mapFromQuery(queryMap);
+		FilterInformationTypeRequest filterRequest = new FilterInformationTypeRequest().mapFromQuery(queryMap);
 
 		assertNull(filterRequest.getSpecification());
 		assertThat(filterRequest.getPageable(), is(defaultPageable));
@@ -45,7 +39,7 @@ public class InformationTypeSpecificationTest {
 	public void mapFromQuery_withPageableNumber_willNotBeDefaultPageable() {
 		Map<String, String> queryMap = Map.of("page", "1");
 
-		FilterRequest filterRequest = new FilterRequest().mapFromQuery(queryMap);
+		FilterInformationTypeRequest filterRequest = new FilterInformationTypeRequest().mapFromQuery(queryMap);
 
 		assertNull(filterRequest.getSpecification());
 		assertNotEquals(defaultPageable, filterRequest.getPageable());
@@ -56,7 +50,7 @@ public class InformationTypeSpecificationTest {
 	public void mapFromQuery_withPageable_willBePage3WithPageSize25() {
 		Map<String, String> queryMap = Map.of("page", "3", "pageSize", "25");
 
-		FilterRequest filterRequest = new FilterRequest().mapFromQuery(queryMap);
+		FilterInformationTypeRequest filterRequest = new FilterInformationTypeRequest().mapFromQuery(queryMap);
 
 		assertNull(filterRequest.getSpecification());
 		assertNotEquals(defaultPageable, filterRequest.getPageable());
@@ -68,7 +62,7 @@ public class InformationTypeSpecificationTest {
 	public void mapFromQuery_withSortById_willHaveSortInPageable() {
 		Map<String, String> queryMap = Map.of("sort", "id");
 
-		FilterRequest filterRequest = new FilterRequest().mapFromQuery(queryMap);
+		FilterInformationTypeRequest filterRequest = new FilterInformationTypeRequest().mapFromQuery(queryMap);
 
 		assertNull(filterRequest.getSpecification());
 		assertNotEquals(defaultPageable, filterRequest.getPageable());
@@ -81,7 +75,7 @@ public class InformationTypeSpecificationTest {
 	public void mapFromQuery_withSortByIdDescending_willHaveSortInPageable() {
 		Map<String, String> queryMap = Map.of("sort", "id,desc");
 
-		FilterRequest filterRequest = new FilterRequest().mapFromQuery(queryMap);
+		FilterInformationTypeRequest filterRequest = new FilterInformationTypeRequest().mapFromQuery(queryMap);
 
 		assertNull(filterRequest.getSpecification());
 		assertNotEquals(defaultPageable, filterRequest.getPageable());
@@ -94,7 +88,7 @@ public class InformationTypeSpecificationTest {
 	public void mapFromQuery_withName_willHaveSpecificationHasName() {
 		Map<String, String> queryMap = Map.of("name", "Sivilstand");
 
-		FilterRequest filterRequest = new FilterRequest().mapFromQuery(queryMap);
+		FilterInformationTypeRequest filterRequest = new FilterInformationTypeRequest().mapFromQuery(queryMap);
 
 		assertNotNull(filterRequest.getSpecification());
 		assertThat(filterRequest.getSpecification(), instanceOf(Specification.class));
